requirements.txt
fastapi
uvicorn
pydantic
torch
transformers
sentence-transformers
scikit-learn
numpy
accelerate
bitsandbytes
psycopg2-binary
python-dotenv